skripte/requirements.txt
click==8.1.3
matplotlib==3.7.1
opencv-python==4.7.0.72
openslide-python==1.2.0
pandas==1.5.3
Pillow==9.4.0
requests==2.28.2
scikit-image==0.20.0
shapely==2.0.1
tqdm==4.65.0
segmentation-models-pytorch==0.3.2
# tribun_patchio==0.0.2
# tribvn_ai==0.0.2
tensorboard==2.13.0
albumentations==1.3.0
# torch==2.0.0+cu117